Add indicator api tests
